🚚 move apps into their own folder
