+++ modules/contracts/contracts/apps/HighRollerApp.sol
