+++ modules/contracts/test/apps/simple-linked-transfer-app.spec.ts
@@ -5,7 +5,7 @@ import { Contract } from "ethers";
 import { AddressZero, Zero } from "ethers/constants";
 import { BigNumber, defaultAbiCoder, solidityKeccak256 } from "ethers/utils";
 
-import SimpleLinkedTransferApp from "../build/SimpleLinkedTransferApp.json";
+import SimpleLinkedTransferApp from "../../build/SimpleLinkedTransferApp.json";
 
 chai.use(waffle.solidity);
